egg: add EggFillBox
EggFillBox is a container that will try to layout children in a series
of columns. It tries to ensure the columns have similar heights. Contrast
this with GtkFlowBox or GtkGrid where columns and rows are similarly sized
with their peers. EggFillBox has no such requirement, other than columns,
which will allocate to the same width.